Moved the remaining routes in the appropriate Blueprints
