+++ enigma/users/routes.py
@@ -2,7 +2,7 @@ from flask import Blueprint
 
 users = Blueprint("users", __name__)
 
-@app.route("/register", methods=["GET", "POST"])
+@users.route("/register", methods=["GET", "POST"])
 def register():
     if current_user.is_authenticated:
         return redirect(url_for("home"))
@@ -16,7 +16,7 @@ def register():
         return redirect(url_for("login"))
     return render_template("register.html", title="Register", form=form)
 
-@app.route("/login", methods=["GET", "POST"])
+@users.route("/login", methods=["GET", "POST"])
 def login():
     if current_user.is_authenticated:
         return redirect(url_for("home"))
@@ -31,12 +31,12 @@ def login():
             flash("Login unsuccessful. Please check email and password.", "danger")
     return render_template("login.html", title="Login", form=form)
 
-@app.route("/logout")
+@users.route("/logout")
 def logout():
     logout_user()
     return redirect(url_for("home"))
 
-@app.route("/account", methods=["GET", "POST"])
+@users.route("/account", methods=["GET", "POST"])
 @login_required
 def account():
     form = UpdateAccountForm()
@@ -55,7 +55,7 @@ def account():
     image_file = url_for("static", filename=f"profile_pics/{current_user.image_file}")
     return render_template("account.html", title="Account", image_file=image_file, form=form)
 
-@app.route("/user/<string:username>")
+@users.route("/user/<string:username>")
 def user_posts(username):
     page = request.args.get("page", 1, type=int)
     user = User.query.filter_by(username=username).first_or_404()
@@ -63,7 +63,7 @@ def user_posts(username):
     return render_template('user_posts.html', posts=posts, user=user)
 
 
-@app.route("/reset_password", methods=["GET", "POST"])
+@users.route("/reset_password", methods=["GET", "POST"])
 def reset_request():
     if current_user.is_authenticated:
         return redirect(url_for("home"))
@@ -75,7 +75,7 @@ def reset_request():
         return redirect(url_for("login"))
     return render_template('reset_request.html', title="Reset Password", form=form)
 
-@app.route("/reset_password/<token>", methods=["GET", "POST"])
+@users.route("/reset_password/<token>", methods=["GET", "POST"])
 def reset_token(token):
     if current_user.is_authenticated:
         return redirect(url_for("home"))
